LPS-106787 Refactor
